+++ src/java/models/CartService.java
@@ -1,12 +1,14 @@
 package models;
 
+import java.math.BigInteger;
+
 public class CartService {
 
     private Cart cart;
     private int cartId;
     private Service service;
     private int quantity;
-    private int priceAtTime;
+    private BigInteger priceAtTime;
 
     public CartService() {
     }
@@ -40,11 +42,11 @@ public class CartService {
     public void setQuantity(int quantity) {
         this.quantity = quantity;
     }
-    
-    public int getPriceAtTime() {
+
+    public BigInteger getPriceAtTime() {
         return priceAtTime;
     }
-    public void setPriceAtTime(int priceAtTime) {
+    public void setPriceAtTime(BigInteger priceAtTime) {
         this.priceAtTime = priceAtTime;
     }
