+++ src/java/models/TypeRoom.java
@@ -1,5 +1,6 @@
 package models;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -8,7 +9,7 @@ public class TypeRoom {
     private int typeId;
     private String typeName;
     private String description;
-    private int price;
+    private BigInteger price;
     private int maxAdult;
     private int maxChildren;
     private List<RoomImage> images = new ArrayList<>();
@@ -18,8 +19,8 @@ public class TypeRoom {
     private double averageRating;
     private List<String> urlImages = new ArrayList<>();
     private int numberOfReviews;
-    private int originPrice;
-    private int servicePrice;
+    private BigInteger originPrice;
+    private BigInteger servicePrice;
     private List<Review> reviews = new ArrayList<>();
     private int adults;
     private int children;
@@ -29,14 +30,14 @@ public class TypeRoom {
     }
 
 
-    public TypeRoom(int typeId, String typeName, String description, int price) {
+    public TypeRoom(int typeId, String typeName, String description, BigInteger price) {
         this.typeId = typeId;
         this.typeName = typeName;
         this.description = description;
         this.price = price;
     }
-    
-    public TypeRoom(int typeId, String typeName, String description, int price, int maxAdult, int maxChildren) {
+
+    public TypeRoom(int typeId, String typeName, String description, BigInteger price, int maxAdult, int maxChildren) {
         this.typeId = typeId;
         this.typeName = typeName;
         this.description = description;
@@ -85,11 +86,11 @@ public class TypeRoom {
         this.description = description;
     }
 
-    public int getPrice() {
+    public BigInteger getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(BigInteger price) {
         this.price = price;
     }
 
@@ -153,16 +154,16 @@ public class TypeRoom {
     public void setNumberOfReviews(int numberOfReviews) {
         this.numberOfReviews = numberOfReviews;
     }
-    public int getOriginPrice() {
+    public BigInteger getOriginPrice() {
         return originPrice;
     }
-    public void setOriginPrice(int originPrice) {
+    public void setOriginPrice(BigInteger originPrice) {
         this.originPrice = originPrice;
     }
-    public int getServicePrice() {
+    public BigInteger getServicePrice() {
         return servicePrice;
     }
-    public void setServicePrice(int servicePrice) {
+    public void setServicePrice(BigInteger servicePrice) {
         this.servicePrice = servicePrice;
     }
     public List<Review> getReviews() {
